+ implement string to glm::vec conversion + fix Deserializer bugs + add array (de)serialization
